refactor: Change the id type from Long to Integer

+++ src/main/java/com/carservice/data/entities/Appointment.java
@@ -3,7 +3,6 @@ package com.carservice.data.entities;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
 import jdk.jfr.BooleanFlag;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
@@ -18,41 +17,41 @@ import java.util.Set;
 @Entity
 @Table(name = "appointment")
 public class Appointment {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "id")
+	private Integer id;
 
-    @ManyToOne(targetEntity = Customer.class)
-    @JoinColumn(name= "customer_email")
-    @NotEmpty(message = "The customer must be set!")
-    private Customer customer;
+	@ManyToOne(targetEntity = Customer.class)
+	@JoinColumn(name = "customer_email")
+	@NotEmpty(message = "The customer must be set!")
+	private Customer customer;
 
-    @ManyToOne(targetEntity = CarCenter.class)
-    @JoinColumn(name = "car_center_id")
-    @NotEmpty(message = "The car center must be set!")
-    private CarCenter carCenter;
+	@ManyToOne(targetEntity = CarCenter.class)
+	@JoinColumn(name = "car_center_id")
+	@NotEmpty(message = "The car center must be set!")
+	private CarCenter carCenter;
 
-    @ManyToOne(targetEntity = Car.class)
-    @JoinColumn(name = "car_license_plate")
-    @NotEmpty(message = "The car must be set!")
-    private Car car;
+	@ManyToOne(targetEntity = Car.class)
+	@JoinColumn(name = "car_license_plate")
+	@NotEmpty(message = "The car must be set!")
+	private Car car;
 
-    @Column(name = "date_created")
-    @NotEmpty(message = "The date created must be set!")
-    private final LocalDate dateCreated = LocalDate.now();
+	@Column(name = "date_created")
+	@NotEmpty(message = "The date created must be set!")
+	private final LocalDate dateCreated = LocalDate.now();
 
-    @Column(name = "date_of_appointment")
-    @NotEmpty(message = "The date of appointment must be set!")
-    @Future(message = "The date of appointment must be in the future!")
-    private LocalDate dateOfAppointment;
+	@Column(name = "date_of_appointment")
+	@NotEmpty(message = "The date of appointment must be set!")
+	@Future(message = "The date of appointment must be in the future!")
+	private LocalDate dateOfAppointment;
 
-    @OneToMany(targetEntity = ServiceJob.class, mappedBy = "appointment")
-    @Column(name = "serviceJobs")
-    @NotEmpty(message = "The service jobs must be set!")
-    private Set<ServiceJob> serviceJobs;
+	@OneToMany(targetEntity = ServiceJob.class, mappedBy = "appointment")
+	@Column(name = "serviceJobs")
+	@NotEmpty(message = "The service jobs must be set!")
+	private Set<ServiceJob> serviceJobs;
 
-    @Column(name = "has_passed")
-    @BooleanFlag
-    private Boolean hasPassed = false;
+	@Column(name = "has_passed")
+	@BooleanFlag
+	private Boolean hasPassed = false;
 }
