+++ src/main/java/com/carservice/dto/CarCenterDTO.java
@@ -12,16 +12,16 @@ import lombok.ToString;
 @AllArgsConstructor
 @ToString
 public class CarCenterDTO {
-    private Long id;
+	private Integer id;
 
-    @NotBlank(message = "The name cannot be empty")
-    @Size(max = 20, message = "The name cannot be longer than {max} characters")
-    private String name;
+	@NotBlank(message = "The name cannot be empty")
+	@Size(max = 20, message = "The name cannot be longer than {max} characters")
+	private String name;
 
-    @Enumerated(EnumType.STRING)
-    private String workWithBrand;
+	@Enumerated(EnumType.STRING)
+	private String workWithBrand;
 
-    private String appointments;
+	private String appointments;
 
-    private String employees;
+	private String employees;
 }
